refactor: Review id 필드명 변경 및 ReportMessageType 패키지 이동
