+++ backend/src/main/java/coursepick/coursepick/infrastructure/mongodb/CourseReader.java
@@ -102,7 +102,7 @@ public class CourseReader implements Converter<Document, Course> {
         for (Document reviewDoc : reviewDocs) {
             Set<String> reportUserIds = new HashSet<>(reviewDoc.getList("reportUserIds", String.class, List.of()));
 
-            String id = reviewDoc.getString("iid");
+            String id = reviewDoc.getString("id");
             String userId = reviewDoc.getString("userId");
             String authorNickname = reviewDoc.getString("authorNickname");
             String content = reviewDoc.getString("content");
