+++ backend/src/main/java/coursepick/coursepick/infrastructure/mongodb/CourseWriter.java
@@ -49,7 +49,7 @@ public class CourseWriter implements Converter<Course, Document> {
         return reviews.stream()
                 .map(review -> {
                     Document reviewDoc = new Document();
-                    reviewDoc.put("iid", review.id());
+                    reviewDoc.put("id", review.id());
                     reviewDoc.put("userId", review.userId());
                     reviewDoc.put("authorNickname", review.authorNickname());
                     reviewDoc.put("content", review.content());
